+++ src/icalendar/tests/prop/test_windows_to_olson_mapping.py
@@ -12,7 +12,7 @@ def test_windows_timezone(tzp):
     """Test that the timezone is mapped correctly to olson."""
     dt = vDatetime.from_ical("20170507T181920", "Eastern Standard Time")
     expected = tzp.localize(datetime(2017, 5, 7, 18, 19, 20), "America/New_York")
-    assert dt.tzinfo == dt.tzinfo
+    assert dt.tzinfo == expected.tzinfo
     assert dt == expected