+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 Minor changes:
 
 - Add a ``weekday`` attribute to ``vWeekday`` components. See `Issue 749 <https://github.com/collective/icalendar/issues/749>`_.
+- Print failure of doctest to aid debugging.
 
 Breaking changes: