+++ CHANGES.rst
@@ -7,7 +7,7 @@ Changelog
 Minor changes:
 
 - Add a ``weekday`` attribute to ``vWeekday`` components. See `Issue 749 <https://github.com/collective/icalendar/issues/749>`_.
-- Add ``vRecur`` example to usage documentation. See `Issue 758 <https://github.com/collective/icalendar/issues/758>`_.
+- Document ``vRecur`` property. See `Issue 758 <https://github.com/collective/icalendar/issues/758>`_.
 
 Breaking changes: