+++ CHANGES.rst
@@ -16,7 +16,7 @@ Breaking changes:
 
 New features:
 
-- ...
+- Add :ref:`Security Policy`
 
 Bug fixes: