+++ CHANGES.rst
@@ -17,7 +17,7 @@ Breaking changes:
 
 New features:
 
-- ...
+- Add :ref:`Security Policy`
 
 Bug fixes: