+++ CHANGES.rst
@@ -18,6 +18,7 @@ Breaking changes:
 
 New features:
 
+- Add :ref:`Security Policy`
 - Python types in documentation now link to their documentation pages using ``intersphinx``.
 
 Bug fixes: